move switch to header
